requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
diffusers==0.24.0
transformers==4.35.2
accelerate==0.24.1
opencv-python==4.8.1.78
insightface==0.7.3
onnxruntime==1.16.3
numpy==1.24.3
safetensors==0.4.0
pydantic==2.5.0
aiofiles==23.2.1
